♻️ refactor : repository join 수정 🎉 add : request.http 파일 추가
