+++ request/board.http
@@ -1,0 +1,59 @@
+###일반 유저 로그인
+POST http://localhost:8080/users/login
+Content-Type: application/json
+
+{
+  "username": "user1",
+  "password": "user1"
+}
+> {% client.global.set("USER_TOKEN", response.headers.valueOf("Authorization"));%}
+
+###매니저 유저 로그인
+POST http://localhost:8080/users/login
+Content-Type: application/json
+
+{
+  "username": "managerUser5",
+  "password": "managerUser1"
+}
+> {% client.global.set("USER_TOKEN", response.headers.valueOf("Authorization")); %}
+
+###보드 생성(매니저만 가능)
+POST http://localhost:8080/boards
+Content-Type: application/json
+Authorization:{{USER_TOKEN}}
+
+{
+  "boardName": "보드4",
+  "boardIntro": "보드 설명"
+}
+
+###보드 수정(매니저만 가능)
+POST http://localhost:8080/boards/1
+Content-Type: application/json
+Authorization:{{USER_TOKEN}}
+
+{
+  "boardName": "보드 수정",
+  "boardIntro": "보드 수정 테스트"
+}
+
+###보드 초대(매니저만 가능)
+POST http://localhost:8080/boards/1/invite/1
+Content-Type: application/x-www-form-urlencoded
+Authorization:{{USER_TOKEN}}
+
+<> 2024-07-15T200349.201.txt
+<> 2024-07-15T194430.500.json
+
+###보드 삭제(매니저만 가능)
+DELETE http://localhost:8080/boards/4
+Content-Type: application/json
+Authorization:{{USER_TOKEN}}
+
+<> 2024-07-15T200406.200.json
+
+###모든 보드 조회
+GET http://localhost:8080/boards
+Content-Type: application/json
+Authorization:{{USER_TOKEN}}
