+++ request/comment.http
@@ -1,0 +1,33 @@
+###일반 유저 로그인
+POST http://localhost:8080/users/login
+Content-Type: application/json
+
+{
+  "username": "user1",
+  "password": "user1"
+}
+> {% client.global.set("USER_TOKEN", response.headers.valueOf("Authorization"));%}
+
+###매니저 유저 로그인
+POST http://localhost:8080/users/login
+Content-Type: application/json
+
+{
+  "username": "managerUser5",
+  "password": "managerUser1"
+}
+> {% client.global.set("USER_TOKEN", response.headers.valueOf("Authorization")); %}
+
+###댓글 작성
+POST http://localhost:8080/cards/1/comments
+Authorization:{{USER_TOKEN}}
+Content-Type: application/json
+
+{
+  "comment": "댓글 1"
+}
+
+###댓글 조회
+GET http://localhost:8080/cards/1/comments
+Authorization:{{USER_TOKEN}}
+Content-Type: application/json
