+++ request/user.http
@@ -1,0 +1,45 @@
+###일반 유저 회원가입
+POST http://localhost:8080/users/signup
+Content-Type: application/json
+
+{
+  "username": "user4",
+  "password": "user1"
+}
+
+###일반 유저 로그인
+POST http://localhost:8080/users/login
+Content-Type: application/json
+
+{
+  "username": "user1",
+  "password": "user1"
+}
+> {% client.global.set("USER_TOKEN", response.headers.valueOf("Authorization"));%}
+
+###매니저 유저 회원가입
+POST http://localhost:8080/users/signup
+Content-Type: application/json
+
+{
+  "username": "managerUser5",
+  "password": "managerUser1",
+  "isManager" : true,
+  "managerToken": "7Ja065Oc66+87YKk"
+}
+
+###매니저 유저 로그인
+POST http://localhost:8080/users/login
+Content-Type: application/json
+
+{
+  "username": "managerUser5",
+  "password": "managerUser1"
+}
+> {% client.global.set("USER_TOKEN", response.headers.valueOf("Authorization")); %}
+
+
+###유저 로그아웃
+POST http://localhost:8080/users/logout
+Content-Type: application/json
+Authorization:{{USER_TOKEN}}
