+++ src/main/java/com/sparta/kanbanboardproject/domain/board/entity/Board.java
@@ -20,7 +20,7 @@ public class Board {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, unique = true)
+    @Column(nullable = false)
     private String boardName;
 
     @Column(nullable = false)
