+++ src/main/java/com/sparta/kanbanboardproject/domain/user/repository/collaborator/CollaboratorRepository.java
@@ -1,4 +1,4 @@
-package com.sparta.kanbanboardproject.domain.user.repository;
+package com.sparta.kanbanboardproject.domain.user.repository.collaborator;
 
 import com.sparta.kanbanboardproject.domain.user.entity.Collaborator;
 import org.springframework.data.jpa.repository.JpaRepository;
