+++ src/main/java/com/sparta/kanbanboardproject/domain/user/repository/collaborator/CollaboratorRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.sparta.kanbanboardproject.domain.user.repository;
+package com.sparta.kanbanboardproject.domain.user.repository.collaborator;
 
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
@@ -18,11 +18,9 @@ public class CollaboratorRepositoryImpl implements CustomCollaboratorRepository 
     @Override
     public boolean existsByUserId(Long invitedUserId) {
         QCollaborator collaborator=QCollaborator.collaborator;
-        QUser user=QUser.user;
 
         Integer fetchFirst = jpaQueryFactory.selectOne()
                 .from(collaborator)
-                .join(collaborator.user, user).fetchJoin()
                 .where(collaborator.user.id.eq(invitedUserId))
                 .fetchFirst();
         return fetchFirst != null;
