+++ src/main/java/com/sparta/kanbanboardproject/domain/user/repository/collaborator/CustomCollaboratorRepository.java
@@ -1,4 +1,4 @@
-package com.sparta.kanbanboardproject.domain.user.repository;
+package com.sparta.kanbanboardproject.domain.user.repository.collaborator;
 
 public interface CustomCollaboratorRepository {
     boolean existsByUserId(Long invitedUserId);
